requirements.txt
python-telegram-bot
gtts
python-dotenv
gpt4all
